pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "turret-analysis"
version = "0.1.0"
description = "Motor telemetry analysis for anti-drone turret system"
readme = "README.md"
requires-python = ">=3.10"
authors = [
    {name = "Turret Analysis Team"}
]
dependencies = [
    "rerun-sdk[datafusion]>=0.18.0",
    "polars>=0.20.0",
    "numpy>=1.26.0",
    "scipy>=1.11.0",
    "matplotlib>=3.8.0",
    "plotly>=5.18.0",
    "jupyter>=1.0.0",
    "notebook>=7.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
]

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_functions = ["test_*"]

[tool.black]
line-length = 100
target-version = ["py310"]

[tool.ruff]
line-length = 100
target-version = "py310"
